Added set_Legend and draw_Leg_Lab functions to improve main's length. Removed nested if in rounding_int function with switch statement. Use of only one try block in all the main.

+++ sir_model/epidemic.hpp
@@ -51,7 +51,7 @@ class Epidemic {
 
   // evolves epidemic by a day
   void evolve();
-  void rounding_int(double tmp_S, double tmp_I, double tmp_R,  double const N);
+  void rounding_int(double tmp_S, double tmp_I, double tmp_R,  int const N);
 };
 
 #endif
